Move tests for each module into separate files
